sanity check

+++ sycl/include/sycl/ext/intel/esimd/detail/util.hpp
@@ -182,7 +182,7 @@ public:
 /// Returns the address referenced by the accessor \p Acc and
 /// the byte offset \p Offset.
 template <typename T, typename AccessorTy>
-T *accessorToPointer(AccessorTy Acc, uint32_t Offset = 0) {
+T *accessorToPointer(AccessorTy Acc, uint64_t Offset = 0) {
   auto BytePtr = reinterpret_cast<char *>(Acc.get_pointer().get()) + Offset;
   return reinterpret_cast<T *>(BytePtr);
 }
